Rename MotorSidebar to MotorSidebarView
